GH-37906: [Integration][C#] Implement C Data Interface integration testing for C#
