+++ csharp/src/Apache.Arrow/C/CArrowArrayExporter.cs
@@ -15,6 +15,7 @@
 
 
 using System;
+using System.Diagnostics;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 using Apache.Arrow.Memory;
@@ -59,8 +60,6 @@ namespace Apache.Arrow.C
             try
             {
                 ConvertArray(allocationOwner, array.Data, cArray);
-                cArray->release = ReleaseArrayPtr;
-                cArray->private_data = FromDisposable(allocationOwner);
                 allocationOwner = null;
             }
             finally
@@ -102,8 +101,6 @@ namespace Apache.Arrow.C
             try
             {
                 ConvertRecordBatch(allocationOwner, batch, cArray);
-                cArray->release = ReleaseArrayPtr;
-                cArray->private_data = FromDisposable(allocationOwner);
                 allocationOwner = null;
             }
             finally
@@ -118,7 +115,7 @@ namespace Apache.Arrow.C
             cArray->offset = array.Offset;
             cArray->null_count = array.NullCount;
             cArray->release = ReleaseArrayPtr;
-            cArray->private_data = null;
+            cArray->private_data = MakePrivateData(sharedOwner);
 
             cArray->n_buffers = array.Buffers?.Length ?? 0;
             cArray->buffers = null;
@@ -131,7 +128,7 @@ namespace Apache.Arrow.C
                     IntPtr ptr;
                     if (!buffer.TryExport(sharedOwner, out ptr))
                     {
-                        throw new NotSupportedException($"An ArrowArray of type {array.DataType.TypeId} could not be exported");
+                        throw new NotSupportedException($"An ArrowArray of type {array.DataType.TypeId} could not be exported: failed on buffer #{i}");
                     }
                     cArray->buffers[i] = (byte*)ptr;
                 }
@@ -144,7 +141,7 @@ namespace Apache.Arrow.C
                 cArray->children = (CArrowArray**)sharedOwner.Allocate(IntPtr.Size * array.Children.Length);
                 for (int i = 0; i < array.Children.Length; i++)
                 {
-                    cArray->children[i] = CArrowArray.Create();
+                    cArray->children[i] = MakeArray(sharedOwner);
                     ConvertArray(sharedOwner, array.Children[i], cArray->children[i]);
                 }
             }
@@ -152,7 +149,7 @@ namespace Apache.Arrow.C
             cArray->dictionary = null;
             if (array.Dictionary != null)
             {
-                cArray->dictionary = CArrowArray.Create();
+                cArray->dictionary = MakeArray(sharedOwner);
                 ConvertArray(sharedOwner, array.Dictionary, cArray->dictionary);
             }
         }
@@ -163,20 +160,24 @@ namespace Apache.Arrow.C
             cArray->offset = 0;
             cArray->null_count = 0;
             cArray->release = ReleaseArrayPtr;
-            cArray->private_data = null;
+            cArray->private_data = MakePrivateData(sharedOwner);
 
             cArray->n_buffers = 1;
             cArray->buffers = (byte**)sharedOwner.Allocate(IntPtr.Size);
 
             cArray->n_children = batch.ColumnCount;
             cArray->children = null;
+            // XXX sharing the same ExportedAllocationOwner for all columns
+            // and child arrays makes memory tracking inflexible.
+            // If the consumer keeps only a single record batch column,
+            // the entire record batch memory is nevertheless kept alive.
             if (cArray->n_children > 0)
             {
                 cArray->children = (CArrowArray**)sharedOwner.Allocate(IntPtr.Size * batch.ColumnCount);
                 int i = 0;
                 foreach (IArrowArray child in batch.Arrays)
                 {
-                    cArray->children[i] = CArrowArray.Create();
+                    cArray->children[i] = MakeArray(sharedOwner);
                     ConvertArray(sharedOwner, child.Data, cArray->children[i]);
                     i++;
                 }
@@ -190,26 +191,43 @@ namespace Apache.Arrow.C
 #endif
         private unsafe static void ReleaseArray(CArrowArray* cArray)
         {
-            Dispose(&cArray->private_data);
+            for (long i = 0; i < cArray->n_children; i++)
+            {
+                CArrowArray.CallReleaseFunc(cArray->children[i]);
+            }
+            if (cArray->dictionary != null) {
+                CArrowArray.CallReleaseFunc(cArray->dictionary);
+            }
+            DisposePrivateData(&cArray->private_data);
             cArray->release = default;
         }
 
-        private unsafe static void* FromDisposable(IDisposable disposable)
+        private unsafe static CArrowArray* MakeArray(ExportedAllocationOwner sharedOwner)
         {
-            GCHandle gch = GCHandle.Alloc(disposable);
+            var array = (CArrowArray*)sharedOwner.Allocate(sizeof(CArrowArray));
+            *array = default;
+            return array;
+        }
+
+        private unsafe static void* MakePrivateData(ExportedAllocationOwner sharedOwner)
+        {
+            GCHandle gch = GCHandle.Alloc(sharedOwner);
+            sharedOwner.IncRef();
             return (void*)GCHandle.ToIntPtr(gch);
         }
 
-        private unsafe static void Dispose(void** ptr)
+        private unsafe static void DisposePrivateData(void** ptr)
         {
-            GCHandle gch = GCHandle.FromIntPtr((IntPtr)(*ptr));
+            GCHandle gch = GCHandle.FromIntPtr((IntPtr) (*ptr));
             if (!gch.IsAllocated)
             {
                 return;
             }
-            ((IDisposable)gch.Target).Dispose();
+            // We can't call IDisposable.Dispose() here as we create multiple
+            // GCHandles to the same object. Instead, refcounting ensures
+            // timely memory deallocation when all GCHandles are freed.
+            ((ExportedAllocationOwner) gch.Target).DecRef();
             gch.Free();
-            *ptr = null;
         }
     }
 }
